renamed fixture verticalnoresults.ts to alternativenoresults.ts
